Renamed angle => rate
