+++ src/pidcontrollers/yaw.hpp
@@ -23,7 +23,7 @@
 #include "receiver.hpp"
 #include "filters.hpp"
 #include "datatypes.hpp"
-#include "angle.hpp"
+#include "rate.hpp"
 #include "pidcontroller.hpp"
 
 namespace hf {
@@ -37,14 +37,14 @@ namespace hf {
         // Arbitrary
         const float BIG_YAW_DEMAND = 0.1f;
 
-        AnglePid _anglePid;
+        RatePid _ratePid;
 
         public:
 
         YawPid(float P, float I, float demandScale = 1.0f, float D=0.0f) 
         {
             // PI controller
-            _anglePid.init(P, I, D, demandScale);
+            _ratePid.init(P, I, D, demandScale);
         }
 
         bool modifyDemands(state_t & state, demands_t & demands, float currentTime)
@@ -56,10 +56,10 @@ namespace hf {
             // Reset integral on large yaw command
             if (fabs(demands.yaw) > BIG_YAW_DEMAND) {
                 itermFactor = 0.0;
-                _anglePid.resetIntegral();
+                _ratePid.resetIntegral();
             }
 
-            demands.yaw = _anglePid.compute(demands.yaw, state.angularVel[2], itermFactor);
+            demands.yaw = _ratePid.compute(demands.yaw, state.angularVel[2], itermFactor);
 
             // Prevent "yaw jump" during gyroYawPid correction
             demands.yaw = Filter::constrainAbs(demands.yaw, 0.1 + fabs(demands.yaw));
@@ -70,7 +70,7 @@ namespace hf {
 
         virtual void updateReceiver(demands_t & demands, bool throttleIsDown) override
         {
-            _anglePid.updateReceiver(demands, throttleIsDown);
+            _ratePid.updateReceiver(demands, throttleIsDown);
         }
 
     };  // class YawPid
